(rcm) update TUF to use int64, fix CVE-2022-29173  (#11972)
* update all tuf imports

* remove int64()

* inv generate-licenses

+++ pkg/remoteconfig/client/internal/uptane/partial_client.go
@@ -200,7 +200,7 @@ func (c *PartialClient) Update(roots [][]byte, previousTargets *PartialClientTar
 	}
 	purgeTargetFiles(targets.Targets, mergedTargetFiles)
 	return &PartialClientTargets{
-		version:     int64(targets.Version),
+		version:     targets.Version,
 		metas:       targets.Targets,
 		targetFiles: mergedTargetFiles,
 	}, nil
